+++ x/ccv/provider/keeper/permissionless_test.go
@@ -4,11 +4,8 @@ import (
 	"testing"
 	"time"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	govv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	"github.com/cosmos/ibc-go/v8/modules/core/02-client/types"
 	testkeeper "github.com/cosmos/interchain-security/v5/testutil/keeper"
-	"github.com/cosmos/interchain-security/v5/x/ccv/provider/keeper"
 	providertypes "github.com/cosmos/interchain-security/v5/x/ccv/provider/types"
 	"github.com/stretchr/testify/require"
 )
@@ -312,128 +309,4 @@ func TestIsValidatorOptedInToChain(t *testing.T) {
 	actualConsumerId, found := providerKeeper.IsValidatorOptedInToChainId(ctx, providerAddr, chainId)
 	require.True(t, found)
 	require.Equal(t, expectedConsumerId, actualConsumerId)
-}
-
-func TestPrepareConsumerForLaunch(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	spawnTime := time.Now().UTC()
-	err := providerKeeper.PrepareConsumerForLaunch(ctx, "consumerId", time.Time{}, spawnTime)
-	require.NoError(t, err)
-
-	consumers, err := providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, providertypes.ConsumerIds{Ids: []string{"consumerId"}}, consumers)
-
-	nextSpawnTime := spawnTime.Add(time.Hour)
-	err = providerKeeper.PrepareConsumerForLaunch(ctx, "consumerId", spawnTime, nextSpawnTime)
-	require.NoError(t, err)
-
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Empty(t, consumers)
-
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, nextSpawnTime)
-	require.NoError(t, err)
-	require.Equal(t, providertypes.ConsumerIds{Ids: []string{"consumerId"}}, consumers)
-}
-
-func TestCanLaunch(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	// cannot launch an unknown chain
-	_, canLaunch := providerKeeper.CanLaunch(ctx, "consumerId")
-	require.False(t, canLaunch)
-
-	// cannot launch a chain without initialization parameters
-	providerKeeper.SetConsumerPhase(ctx, "consumerId", providertypes.ConsumerPhase_CONSUMER_PHASE_INITIALIZED)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.False(t, canLaunch)
-
-	// set valid initialization parameters
-	initializationParameters := testkeeper.GetTestInitializationParameters()
-	err := providerKeeper.SetConsumerInitializationParameters(ctx, "consumerId", initializationParameters)
-	require.NoError(t, err)
-
-	// cannot launch a launched chain
-	providerKeeper.SetConsumerPhase(ctx, "consumerId", providertypes.ConsumerPhase_CONSUMER_PHASE_LAUNCHED)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.False(t, canLaunch)
-
-	// cannot launch a stopped chain
-	providerKeeper.SetConsumerPhase(ctx, "consumerId", providertypes.ConsumerPhase_CONSUMER_PHASE_STOPPED)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.False(t, canLaunch)
-
-	// initialized chain can launch
-	providerKeeper.SetConsumerPhase(ctx, "consumerId", providertypes.ConsumerPhase_CONSUMER_PHASE_INITIALIZED)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.True(t, canLaunch)
-
-	// chain cannot launch without a genesis hash
-	initializationParameters.GenesisHash = nil
-	err = providerKeeper.SetConsumerInitializationParameters(ctx, "consumerId", initializationParameters)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.NoError(t, err)
-	require.False(t, canLaunch)
-}
-
-func TestHasAtMostOnceCorrectMsgUpdateConsumer(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	expectedMsgUpdateConsumer := providertypes.MsgUpdateConsumer{Signer: "signer", ConsumerId: "consumerId", NewOwnerAddress: "new owner address"}
-
-	proposal, err := govv1.NewProposal([]sdk.Msg{&expectedMsgUpdateConsumer}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-
-	_, err = providerKeeper.HasAtMostOnceCorrectMsgUpdateConsumer(ctx, &proposal)
-	require.ErrorContains(t, err, "cannot find owner address")
-
-	// set owner address that is not the gov module
-	providerKeeper.SetConsumerOwnerAddress(ctx, "consumerId", "owner address")
-	_, err = providerKeeper.HasAtMostOnceCorrectMsgUpdateConsumer(ctx, &proposal)
-	require.ErrorContains(t, err, "is not the gov module")
-
-	// set owner address that is the gov module
-	providerKeeper.SetConsumerOwnerAddress(ctx, "consumerId", "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn")
-	actualMsgUpdateConsumer, err := providerKeeper.HasAtMostOnceCorrectMsgUpdateConsumer(ctx, &proposal)
-	require.NoError(t, err)
-	require.Equal(t, expectedMsgUpdateConsumer, *actualMsgUpdateConsumer)
-
-	// a proposal with 2 `MsgUpdateConsumer` messages
-	invalidProposal, err := govv1.NewProposal([]sdk.Msg{&expectedMsgUpdateConsumer, &expectedMsgUpdateConsumer}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	actualMsgUpdateConsumer, err = providerKeeper.HasAtMostOnceCorrectMsgUpdateConsumer(ctx, &invalidProposal)
-	require.ErrorContains(t, err, "proposal can contain at most one")
-	require.Nil(t, actualMsgUpdateConsumer)
-}
-
-func TestDoesNotHaveDeprecatedMessage(t *testing.T) {
-	msgConsumerAddition := providertypes.MsgConsumerAddition{}
-	proposal, err := govv1.NewProposal([]sdk.Msg{&msgConsumerAddition}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	err = keeper.DoesNotHaveDeprecatedMessage(&proposal)
-	require.ErrorContains(t, err, "cannot contain deprecated `MsgConsumerAddition`")
-
-	msgConsumerModification := providertypes.MsgConsumerModification{}
-	proposal, err = govv1.NewProposal([]sdk.Msg{&msgConsumerModification}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	err = keeper.DoesNotHaveDeprecatedMessage(&proposal)
-	require.ErrorContains(t, err, "cannot contain deprecated `MsgConsumerModification`")
-
-	msgConsumerRemoval := providertypes.MsgConsumerRemoval{}
-	proposal, err = govv1.NewProposal([]sdk.Msg{&msgConsumerRemoval}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	err = keeper.DoesNotHaveDeprecatedMessage(&proposal)
-	require.ErrorContains(t, err, "cannot contain deprecated `MsgConsumerRemoval`")
-
-	// a proposal with no deprecated messages
-	msgUpdateConsumer := providertypes.MsgUpdateConsumer{Signer: "signer", ConsumerId: "consumerId", NewOwnerAddress: "new owner address"}
-	proposal, err = govv1.NewProposal([]sdk.Msg{&msgUpdateConsumer}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	err = keeper.DoesNotHaveDeprecatedMessage(&proposal)
-	require.Nil(t, err)
 }
